Backend/image-classification/requirements.txt
Flask==3.0.2
Flask_Cors==4.0.0
Pillow==10.2.0
torch==2.2.0
transformers==4.38.2
